+++ R/denominator.R
@@ -8,7 +8,8 @@
 #'     If `TRUE`, the denominator of `4` will be "quarter(s)".
 #'     If `FALSE`, the denominator of `4` will be "fourth(s)".
 #'     Defaults to `TRUE`.
-#'     Default can be changed by setting `options("nombre.quarter")`.
+#'     Default can be changed with
+#'     [set_config("nombre::quarter")][set_config()].
 #' @param ... Additional arguments passed to [ordinal()]
 #'
 #' @return A character vector of the same length as `x`
@@ -17,7 +18,7 @@
 #' @example examples/denominator.R
 
 denominator <- function(
-  x, numerator = 1, quarter = getOption("nombre.quarter", TRUE), ...
+  x, numerator = 1, quarter = get_config("nombre::quarter", TRUE), ...
 ) {
   if (!length(x))              return(character(0))
   if (!is.numeric(x))          stop("`x` must be numeric")
@@ -28,8 +29,9 @@ denominator <- function(
   if (!is.logical(quarter) | is.na(quarter))
     stop("`quarter` must be either `TRUE` or `FALSE`")
 
-  denom  <- ordinal(x, ...)
-  plural <- abs(numerator) != 1
+  numeric <- x
+  denom   <- ordinal(x, ...)
+  plural  <- abs(numerator) != 1
 
   denom[abs(x) == 1] <- gsub("1st$|first$", "whole", denom[abs(x) == 1])
 
@@ -40,12 +42,24 @@ denominator <- function(
     "2nd$|second$", "halves", denom[abs(x) == 2 & plural]
   )
 
-  if (quarter)
+  if (quarter) {
     denom[abs(x) == 4] <- gsub("fourth$", "quarter", denom[abs(x) == 4])
+  }
 
   denom[plural & abs(x) != 2] <- paste0(denom[plural & abs(x) != 2], "s")
 
-  denom
+  denom <- gsub("^one-", "", denom)
+
+  args        <- as.list(match.call()[-1])
+  args[["x"]] <- NULL
+
+  structure(
+    denom,
+    numeric = numeric,
+    nombre  = "denominator",
+    args    = args,
+    class   = c("nombre", "character")
+  )
 }
 
 #' @rdname denominator